准备开发消息管理机制

+++ LobinaryPlatform/src/main/java/com/lobinary/platform/dao/MessageDAO.java
@@ -3,7 +3,7 @@ package com.lobinary.platform.dao;
 import org.springframework.stereotype.Component;
 
 import com.lobinary.platform.model.DatabaseParameter;
-import com.lobinary.platform.model.ReceiverMessage;
+import com.lobinary.platform.model.db.ReceiverMessage;
 import com.lobinary.platform.util.LogUtil;
 
 
@@ -18,7 +18,7 @@ import com.lobinary.platform.util.LogUtil;
 @Component("messageDAO")
 public class MessageDAO extends BaseDAO {
 
-	public void deleteMessageByID(int id) {
+	public void deleteMessageByID(long id) {
 		try {
 			delete(ReceiverMessage.class, new DatabaseParameter("id", id));
 		} catch (Exception e) {
